Create stuckbreak.py
Script for choosing an item out of a list provided by the user. The list can be provided in the form of a csv file or from the user manually inputting items when the script is run (using "-t 2" upon execution). The items in the list are "spun" where one item is eventually selected.